# metaclouds-backend-rust/docs/openapi-rust.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Metaclouds Backend API",
    "description": "Rust 版 Metaclouds AI 算力调度后端，对齐 Go v1 API 契约。\n\n统一响应信封：`{\"success\":bool,\"data\":...,\"message\":...,\"code\":...,\"timestamp\":sec}`。错误响应对见 `ErrorResponse`。",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8001",
      "description": "development"
    }
  ],
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "`GET /api/v1/jobs` — 分页列表（搜索 + status/type/cluster_id/user_id 过滤）。",
        "operationId": "list_jobs",
        "responses": {
          "200": {
            "description": "paginated jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPage"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "`POST /api/v1/jobs` — 创建（201）。",
        "operationId": "create_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/jobs/stats": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "`GET /api/v1/jobs/stats` — 按状态统计数量。",
        "operationId": "get_job_stats",
        "responses": {
          "200": {
            "description": "job stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "propertyNames": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/jobs/{id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "`GET /api/v1/jobs/:id` — 详情。",
        "operationId": "get_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "jobs"
        ],
        "summary": "`PUT /api/v1/jobs/:id` — 更新（状态机校验）。",
        "operationId": "update_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "`DELETE /api/v1/jobs/:id` — 软删除，204。",
        "operationId": "delete_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "deleted"
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/jobs/{id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "`POST /api/v1/jobs/:id/cancel` — 取消（仅 pending/running）。",
        "operationId": "cancel_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateJobRequest": {
        "type": "object",
        "description": "`POST /api/v1/jobs` 请求体。",
        "required": [
          "name"
        ],
        "properties": {
          "cluster_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "cpus": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "gpus": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "memory": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "统一错误响应 schema（对齐 `ApiResponse` 错误形式）。\n\n实际信封：`{\"success\":false,\"message\":\"...\",\"code\":\"...\",\"timestamp\":<sec>}`。\n成功响应的 `data` 字段由各 handler 对应的业务 schema 描述。",
        "required": [
          "success",
          "message",
          "code",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "JobPage": {
        "type": "object",
        "description": "分页作业列表响应内层。",
        "required": [
          "data",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "page_size": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "total_pages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "JobResponse": {
        "type": "object",
        "description": "对外作业视图（剔除软删除列 `deleted_at`，对齐 Go `json:\"-\"`）。",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "cluster_id",
          "tenant_id",
          "user_id",
          "name",
          "description",
          "status",
          "type",
          "priority",
          "gpus",
          "cpus",
          "memory",
          "duration",
          "progress",
          "output_path",
          "error_msg",
          "partition_id",
          "scheduler_type",
          "scheduler_job_id",
          "gpu_fraction",
          "gpu_memory_gb",
          "gpu_vendor",
          "qos",
          "nodes_requested",
          "node_selector",
          "affinity",
          "tolerations",
          "elastic_enabled",
          "min_gpus",
          "max_gpus",
          "scaling_policy",
          "checkpoint_enabled",
          "checkpoint_interval_minutes",
          "max_retries",
          "retry_count",
          "fault_tolerance_level",
          "topology_affinity",
          "topology_anti_affinity",
          "network_requirement"
        ],
        "properties": {
          "affinity": {
            "type": "string"
          },
          "checkpoint_enabled": {
            "type": "boolean"
          },
          "checkpoint_interval_minutes": {
            "type": "integer",
            "format": "int64"
          },
          "cluster_id": {
            "type": "integer",
            "format": "int64"
          },
          "cpus": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "format": "int64"
          },
          "elastic_enabled": {
            "type": "boolean"
          },
          "end_time": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "error_msg": {
            "type": "string"
          },
          "fault_tolerance_level": {
            "type": "string"
          },
          "gpu_fraction": {
            "type": "number",
            "format": "double"
          },
          "gpu_memory_gb": {
            "type": "integer",
            "format": "int64"
          },
          "gpu_vendor": {
            "type": "string"
          },
          "gpus": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "max_gpus": {
            "type": "integer",
            "format": "int64"
          },
          "max_retries": {
            "type": "integer",
            "format": "int64"
          },
          "memory": {
            "type": "integer",
            "format": "int64"
          },
          "min_gpus": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "network_requirement": {
            "type": "string"
          },
          "node_selector": {
            "type": "string"
          },
          "nodes_requested": {
            "type": "integer",
            "format": "int64"
          },
          "output_path": {
            "type": "string"
          },
          "partition_id": {
            "type": "integer",
            "format": "int64"
          },
          "priority": {
            "type": "integer",
            "format": "int64"
          },
          "progress": {
            "type": "integer",
            "format": "int64"
          },
          "qos": {
            "type": "string"
          },
          "retry_count": {
            "type": "integer",
            "format": "int64"
          },
          "scaling_policy": {
            "type": "string"
          },
          "scheduler_job_id": {
            "type": "string"
          },
          "scheduler_type": {
            "type": "string"
          },
          "start_time": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "tenant_id": {
            "type": "integer",
            "format": "int64"
          },
          "tolerations": {
            "type": "string"
          },
          "topology_affinity": {
            "type": "string"
          },
          "topology_anti_affinity": {
            "type": "boolean"
          },
          "type": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "UpdateJobRequest": {
        "type": "object",
        "description": "`PUT /api/v1/jobs/:id` 请求体（全部可选）。",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "error_msg": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "output_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "priority": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "progress": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookie_auth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  },
...
}